Refactor: Object Loading

+++ Modules/RenderCore/Source/Utils/VulkanConstants.h
@@ -59,7 +59,7 @@ namespace RenderCore
         VK_DYNAMIC_STATE_SCISSOR
     };
 
-    constexpr std::uint32_t g_MaxFramesInFlight = 2u;
+    constexpr std::uint8_t g_MaxFramesInFlight = 2u;
 
     constexpr std::uint32_t g_BufferMemoryAllocationSize = 65536u;
 }
